Fixed the problem with cyclical links.
Now, the run_tests.sh script accepts symlink definition files, creates the symlinks when
tests are run and cleans them up when the --clean option is passed.

Created helpers.sh, which is a helper script that runs functions associated with the
symlinks. Required as a separate script because xargs can't run bash builtins.


git-svn-id: https://cxxtest.svn.sourceforge.net/svnroot/cxxtest/trunk@102 2f8b185b-e3eb-40a7-945f-4de83c84e57e